small tweaks and reorganize session queries
