+++ ee/clickhouse/models/person.py
@@ -1,7 +1,6 @@
 import datetime
 import json
 from typing import Any, Dict, List, Optional
-from uuid import UUID
 
 from django.db.models.query import QuerySet
 from django.db.models.signals import post_delete, post_save
@@ -17,16 +16,14 @@ from ee.clickhouse.sql.person import (
     GET_DISTINCT_IDS_SQL_BY_ID,
     GET_PERSON_BY_DISTINCT_ID,
     GET_PERSON_SQL,
-    GET_PERSONS_BY_DISTINCT_IDS,
     INSERT_PERSON_DISTINCT_ID,
     INSERT_PERSON_SQL,
     PERSON_DISTINCT_ID_EXISTS_SQL,
-    PERSON_EXISTS_SQL,
     UPDATE_PERSON_ATTACHED_DISTINCT_ID,
     UPDATE_PERSON_IS_IDENTIFIED,
     UPDATE_PERSON_PROPERTIES,
 )
-from ee.kafka_client.client import ClickhouseProducer, KafkaProducer
+from ee.kafka_client.client import ClickhouseProducer
 from ee.kafka_client.topics import KAFKA_PERSON, KAFKA_PERSON_UNIQUE_ID
 from posthog import settings
 from posthog.ee import check_ee_enabled
@@ -98,10 +95,6 @@ def create_person_distinct_id(id: int, team_id: int, distinct_id: str, person_id
 
 def distinct_ids_exist(team_id: int, ids: List[str]) -> bool:
     return bool(sync_execute(PERSON_DISTINCT_ID_EXISTS_SQL.format([str(id) for id in ids]), {"team_id": team_id})[0][0])
-
-
-def person_exists(id: int) -> bool:
-    return bool(sync_execute(PERSON_EXISTS_SQL, {"id": id})[0][0])
 
 
 def get_persons(team_id: int):
