+++ ee/clickhouse/queries/clickhouse_sessions.py
@@ -8,209 +8,18 @@ from ee.clickhouse.models.event import ClickhouseEventSerializer
 from ee.clickhouse.models.person import get_persons_by_distinct_ids
 from ee.clickhouse.models.property import parse_prop_clauses
 from ee.clickhouse.queries.util import get_interval_annotation_ch, get_time_diff, parse_timestamps
-from ee.clickhouse.sql.events import GET_EARLIEST_TIMESTAMP_SQL, NULL_SQL
+from ee.clickhouse.sql.events import NULL_SQL
+from ee.clickhouse.sql.sessions.average_all import AVERAGE_SQL
+from ee.clickhouse.sql.sessions.average_per_period import AVERAGE_PER_PERIOD_SQL
+from ee.clickhouse.sql.sessions.distribution import DIST_SQL
+from ee.clickhouse.sql.sessions.list import SESSION_SQL
+from ee.clickhouse.sql.sessions.no_events import SESSIONS_NO_EVENTS_SQL
 from posthog.constants import SESSION_AVG, SESSION_DIST
 from posthog.models import Filter, Person, Team
 from posthog.queries.base import BaseQuery, determine_compared_filter
 from posthog.utils import append_data, friendly_time
 
 SESSIONS_LIST_DEFAULT_LIMIT = 50
-
-SESSIONS_NO_EVENTS_SQL = """
-SELECT
-    distinct_id,
-    uuid,
-    event,
-    properties,
-    elements_chain,
-    session_uuid,
-    session_duration_seconds,
-    timestamp,
-    session_end_ts
-FROM
-(
-    SELECT
-        distinct_id,
-        uuid,
-        event,
-        properties,
-        elements_chain,
-        if(is_new_session, uuid, NULL) AS session_uuid,
-        is_new_session,
-        is_end_session,
-        if(is_end_session AND is_new_session, 0, if(is_new_session AND (NOT is_end_session), dateDiff('second', toDateTime(timestamp), toDateTime(neighbor(timestamp, 1))), NULL)) AS session_duration_seconds,
-        timestamp,
-        if(is_end_session AND is_new_session, timestamp, if(is_new_session AND (NOT is_end_session), neighbor(timestamp, 1), NULL)) AS session_end_ts
-    FROM
-    (
-        SELECT
-            distinct_id,
-            uuid,
-            event,
-            properties,
-            elements_chain,
-            timestamp,
-            neighbor(distinct_id, -1) AS start_possible_neighbor,
-            neighbor(timestamp, -1) AS start_possible_prev_ts,
-            if((start_possible_neighbor != distinct_id) OR (dateDiff('minute', toDateTime(start_possible_prev_ts), toDateTime(timestamp)) > 30), 1, 0) AS is_new_session,
-            neighbor(distinct_id, 1) AS end_possible_neighbor,
-            neighbor(timestamp, 1) AS end_possible_prev_ts,
-            if((end_possible_neighbor != distinct_id) OR (dateDiff('minute', toDateTime(timestamp), toDateTime(end_possible_prev_ts)) > 30), 1, 0) AS is_end_session
-        FROM
-        (
-            SELECT
-                uuid,
-                event,
-                properties,
-                timestamp,
-                distinct_id,
-                elements_chain
-            FROM events
-            WHERE 
-                team_id = %(team_id)s
-                {date_from}
-                {date_to} 
-                {filters}
-            GROUP BY
-                uuid,
-                event,
-                properties,
-                timestamp,
-                distinct_id,
-                elements_chain
-            ORDER BY
-                distinct_id ASC,
-                timestamp ASC
-        )
-    )
-    WHERE (is_new_session AND (NOT is_end_session)) OR (is_end_session AND (NOT is_new_session)) OR (is_end_session AND is_new_session)
-)
-WHERE is_new_session
-{sessions_limit}
-"""
-
-SESSION_SQL = """
-    SELECT 
-        distinct_id, 
-        gid, 
-        dateDiff('second', toDateTime(arrayReduce('min', groupArray(timestamp))), toDateTime(arrayReduce('max', groupArray(timestamp)))) AS elapsed,
-        arrayReduce('min', groupArray(timestamp)) as start_time,
-        groupArray(uuid) uuids, 
-        groupArray(event) events, 
-        groupArray(properties) properties, 
-        groupArray(timestamp) timestamps, 
-        groupArray(elements_chain) elements_chain
-    FROM (
-        SELECT
-            distinct_id, 
-            event,
-            timestamp,
-            uuid,
-            properties,
-            elements_chain,
-            arraySum(arraySlice(gids, 1, idx)) AS gid
-        FROM (
-            SELECT 
-                groupArray(timestamp) as timestamps, 
-                groupArray(event) as events, 
-                groupArray(uuid) as uuids, 
-                groupArray(properties) as property_list, 
-                groupArray(elements_chain) as elements_chains, 
-                groupArray(distinct_id) as distinct_ids, 
-                groupArray(new_session) AS gids
-            FROM (
-                SELECT 
-                    distinct_id,
-                    uuid,
-                    event,
-                    properties,
-                    elements_chain,
-                    timestamp, 
-                    neighbor(distinct_id, -1) as possible_neighbor,
-                    neighbor(timestamp, -1) as possible_prev, 
-                    if(possible_neighbor != distinct_id or dateDiff('minute', toDateTime(possible_prev), toDateTime(timestamp)) > 30, 1, 0) as new_session
-                FROM (
-                    SELECT 
-                        uuid,
-                        event,
-                        properties,
-                        timestamp, 
-                        distinct_id,
-                        elements_chain
-                    FROM    
-                        events 
-                    WHERE 
-                        team_id = %(team_id)s
-                        {date_from}
-                        {date_to} 
-                        {filters}
-                    GROUP BY 
-                        uuid,
-                        event,
-                        properties,
-                        timestamp, 
-                        distinct_id,
-                        elements_chain 
-                    ORDER BY 
-                        distinct_id, 
-                        timestamp
-                )
-            )
-        )
-        ARRAY JOIN
-            distinct_ids as distinct_id,
-            events as event,
-            timestamps as timestamp,
-            uuids as uuid,
-            property_list as properties,
-            elements_chains as elements_chain,
-            arrayEnumerate(gids) AS idx 
-    ) 
-    GROUP BY 
-        distinct_id, 
-        gid
-    {sessions_limit}
-"""
-
-AVERAGE_PER_PERIOD_SQL = """
-    SELECT 
-        AVG(session_duration_seconds) as total, 
-        {interval}(timestamp) as day_start 
-    FROM 
-        ({sessions}) 
-    GROUP BY 
-        {interval}(timestamp)
-"""
-
-AVERAGE_SQL = """
-    SELECT 
-        SUM(total), 
-        day_start 
-    FROM 
-        ({null_sql} UNION ALL {sessions}) 
-    GROUP BY 
-        day_start 
-    ORDER BY 
-        day_start
-"""
-
-DIST_SQL = """
-    SELECT 
-        countIf(session_duration_seconds = 0)  as first,
-        countIf(session_duration_seconds > 0 and session_duration_seconds <= 3)  as second,
-        countIf(session_duration_seconds > 3 and session_duration_seconds <= 10)  as third,
-        countIf(session_duration_seconds > 10 and session_duration_seconds <= 30)  as fourth,
-        countIf(session_duration_seconds > 30 and session_duration_seconds <= 60)  as fifth,
-        countIf(session_duration_seconds > 60 and session_duration_seconds <= 180)  as sixth,
-        countIf(session_duration_seconds > 180 and session_duration_seconds <= 600)  as sevent,
-        countIf(session_duration_seconds > 600 and session_duration_seconds <= 1800)  as eighth,
-        countIf(session_duration_seconds > 1800 and session_duration_seconds <= 3600)  as ninth,
-        countIf(session_duration_seconds > 3600)  as tenth
-    FROM 
-        ({sessions})
-""".format(
-    sessions=SESSIONS_NO_EVENTS_SQL
-)
 
 # TODO: handle date and defaults
 class ClickhouseSessions(BaseQuery):
