+++ ee/clickhouse/sql/sessions/average_all.py
@@ -1,0 +1,11 @@
+AVERAGE_SQL = """
+    SELECT 
+        SUM(total), 
+        day_start 
+    FROM 
+        ({null_sql} UNION ALL {sessions}) 
+    GROUP BY 
+        day_start 
+    ORDER BY 
+        day_start
+"""
